Refactor DownloadTarget and DownloadPart
